refactor: change evm form schema to simple yup object rather than memoised function
